Re #2897

+++ Code/Mantid/Framework/Algorithms/src/FindCenterOfMassPosition.cpp
@@ -107,8 +107,8 @@ void FindCenterOfMassPosition::exec()
   int n_iteration = 0;
 
   // Get the number of monitors. We assume that all monitors are stored in the first spectra
-  int n_monitors = inputWS->getInstrument()->getMonitors().size();
-  const size_t numSpec = inputWS->getNumberHistograms();
+  int n_monitors = static_cast<int>(inputWS->getInstrument()->getMonitors().size());
+  const int numSpec = static_cast<int>(inputWS->getNumberHistograms());
 
   // Find center of mass and iterate until we converge
   // to within a quarter of a pixel
